breaking the main file down
